chp_api/requirements.txt
tqdm
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
bmt
reasoner_pydantic
#reasoner-validator
django-environ
django-hosts
gunicorn
django
requests
requests-cache
django-filter
celery
flower
redis
pandas
